+++ src/main/java/jp/oiyokan/OiyokanConstants.java
@@ -27,7 +27,7 @@ public class OiyokanConstants {
     /**
      * Oiyokan のバージョン番号
      */
-    public static final String VERSION = "1.14.20210510a";
+    public static final String VERSION = "1.14.20210518a";
 
     /**
      * 暗号化で利用するパスフレーズ。環境変数 OIYOKAN_PASSPHRASE で上書き動作。
